refactor: :recycle: move cipher function from challenge10 to secret.js, add type: module to package.json to enable import and export functions
