+++ example.txt.enc
