+++ package.json
@@ -15,5 +15,6 @@
   "bugs": {
     "url": "https://github.com/digital-futures-academy/js-fundamentals/issues"
   },
-  "homepage": "https://github.com/digital-futures-academy/js-fundamentals#readme"
+  "homepage": "https://github.com/digital-futures-academy/js-fundamentals#readme",
+  "type": "module"
 }
